requirements.txt
Django==3.0
djangorestframework==3.11
psycopg2
pyYAML
markdown
django-filter
drf_yasg
